migrate tests to TypeScript
