+++ package.json
@@ -12,7 +12,7 @@
   "private": true,
   "scripts": {
     "fix": "prettier --ignore-path .gitignore --write . !package-lock.json !tests/fixtures/**",
-    "test": "tsc -noEmit && NODE_OPTIONS=\"$NODE_OPTIONS --experimental-vm-modules\" jest"
+    "test": "tsc -noEmit && tsx --test $(find tests -name '*.test.ts')"
   },
   "dependencies": {
     "@aws-sdk/client-bedrock-runtime": "3.654.0",
@@ -30,6 +30,7 @@
     "esbuild": "0.23.1",
     "jest": "29.7.0",
     "prettier": "3.3.3",
+    "tsx": "4.19.1",
     "typescript": "5.6.2"
   }
 }
