+++ tests/domain/press-release-rss.test.ts
@@ -1,13 +1,13 @@
-import { expect, test } from "@jest/globals";
+import { test, type TestContext } from "node:test";
 import { retrieveRecallPressReleaseFeedItems } from "../../lib/domain/press-release-rss/index.mjs";
 import { pressreleaseRdf20240908Binary } from "../fixtures.mjs";
 
-test("2024-09-08時点のRSSをparseして「リコールの届出について」のみを古い順に出力すること", async () => {
+test("2024-09-08時点のRSSをparseして「リコールの届出について」のみを古い順に出力すること", async (t: TestContext) => {
   const fn = retrieveRecallPressReleaseFeedItems({
     downloadResource: pressreleaseRdf20240908Binary,
   });
   const actual = await fn();
-  expect(actual).toEqual([
+  t.assert.deepStrictEqual(actual, [
     {
       title: "リコールの届出について（マツダ　MAZDA CX-5　他）",
       link: "http://www.mlit.go.jp/report/press/jidosha08_hh_005220.html",
