+++ tests/domain/recall-press-release-page/utils.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, test } from "@jest/globals";
+import { describe, test, type TestContext } from "node:test";
 import {
   extractPdfLinks,
   extractPreamble,
@@ -6,20 +6,21 @@ import {
 import { recall20240906 } from "../../fixtures.mjs";
 
 describe("extractPreamble", () => {
-  test("2024-09-06のプレスリリースページから冒頭の文章を抽出できること", async () => {
+  test("2024-09-06のプレスリリースページから冒頭の文章を抽出できること", async (t: TestContext) => {
     const dom = await recall20240906();
     const actual = extractPreamble(dom);
-    expect(actual).toBe(
+    t.assert.strictEqual(
+      actual,
       "ドゥカティジャパン株式会社から、令和６年９月６日国土交通大臣に対して、下記のとおりリコールの届出がありましたので、お知らせします。",
     );
   });
 });
 
 describe("extractPdfLinks", () => {
-  test("2024-09-06のプレスリリースページからPDFのリンクを抽出できること", async () => {
+  test("2024-09-06のプレスリリースページからPDFのリンクを抽出できること", async (t: TestContext) => {
     const dom = await recall20240906();
     const actual = extractPdfLinks(dom);
-    expect(actual).toEqual([
+    t.assert.deepStrictEqual(actual, [
       {
         title: "リコール届出一覧表",
         href: "https://www.mlit.go.jp/report/press/content/001761787.pdf",
